Referências

Conteúdo

Referências

Conteúdo

Índice

Índice

Fease.Malha

Fease.elem_coords

Fease.jacobiano

Fease.montaF_geral

Fease.montaFᵉ_geral!

Fease.montaK_geral

Fease.montaKᵉ_geral!

Fease.monta_malha_2D_foco

Fease.monta_u_aproximada

Fease.mudanca_variavel_xξ

Fease.quadratura_gauss

Fease.quadratura_ϕ

Fease.quadratura_∇ϕ

Fease.solve_sys

Fease.solve_sys_poisson

Fease.Malha — Type

struct Mesh

Valores que definem uma malha.

Campos

base::Base: Informações sobre o tipo da base de funções interpoladoras

ne::Int64: Número de elementos totais

neq::Int64: Número de equações

coords::Tuple: N-upla com as coordenadas (X₁, X₂, Xᵢ...) dos nós da malha

dx: n-upla contendo o intervalo entre os nós, para cada eixo (se uniforme)

EQ: Vetor com as reenumerações das funções globais ϕ

LG: Matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ

EQoLG::Matrix{Int64}: Matriz composta entre a EQ e a LG

a: Coordenada do início do intervalo uniforme

b: Coordenada do final do intervalo uniforme

n_dim: Número de dimensões da malha

Nx: Número de subdivisões da malha para cada eixo

Fease.elem_coords — Method

elem_coords(malha::Malha, e::Int)

Encontra as coordenadas dos nós do elemento finito e e os retorna, junto a EQoLG referente.

Parâmetros

malha::Malha: A malha sob a qual está sendo solucionado o problema.

e::Int: O identificador numérico do elemento finito Ωᵉ.

Retorno

eqs_idx::: EQoLG para o elemento e passado como parâmetro

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

Exemplo

Fease.jacobiano — Method

jacobian(n_dim::Int, Xᵉ::Vector, ∇Φξ::Tuple{Matrix{Float64}}, ξ::Int)

Calcula a matriz jacobiana, assim como seu determinante, no domínio do elemento determinado pelas coordenadas em Xᵉ.

Parâmetros

n_dim::Int: Número de dimensões do espaço.

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

∇Φξ::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

ξ::Int: Representa o número do ponto de Gauss.

Retorno

M::Matrix{Float64}: Matriz jacobiana.

detJ::Float64: O determinante da matriz jacobiana

Exemplo

Fease.montaF_geral — Method

montaF_geral(f::Function, malha::Malha)

Monta o vetor F global do sistema linear do problema.

Parâmetros

f::function: Funçãp que represnenta o lado direito da equação do problema.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

F::Vector{Float64}: O vetor F global do sistema linear do problema.

Exemplo

Fease.montaFᵉ_geral! — Method

montaFᵉ_geral!(Fᵉ, f, Xᵉ, P, W, ϕξ, ∇ϕξ, n_dim)

Monta o vetor local Fᵉ referente à contribuição do elemento e.

Parâmetros

Fᵉ::Vector{Float64}: Vetor local Fᵉ referente à contribuição do elemento e.

f::function: Funçãp que represnenta o lado direito da equação do problema.

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

ϕξ::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

∇ϕξ::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço.

Retorno

Altera Fᵉ.

Exemplo

Fease.montaK_geral — Method

montaK_geral(malha::Malha, pseudo_a)

Monta a matriz K global do sistema linear do problema.

Parâmetros

malha::Malha: A malha sob a qual está sendo solucionado o problema.

pseudo_a::: A referência ao operador bilinear a(u,v).

Retorno

K::Matrix{Float64}: A matriz K global do sistema linear do problema.

Exemplo

Fease.montaKᵉ_geral! — Method

montaKᵉ_geral!(Kᵉ, Xᵉ, P, W, Φξ, ∇Φξ, n_dim, pseudo_a)

Monta a matriz local Kᵉ referente à contribuição do elemento e.

Parâmetros

Kᵉ::Matrix{Float64}: Matriz local Kᵉ referente à contribuição do elemento e.

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Φξ::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

∇Φξ::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço.

pseudo_a: A referência ao operador bilinear a(u,v).

Retorno

Altera Kᵉ.

Exemplo

Fease.monta_malha_2D_foco — Function

monta_malha_2D_foco(
baseType, Nx1, Nx2, a::Tuple, b::Tuple, ponto_foco::Tuple, precisao::Int64)::Malha

Função que constrói uma malha 2D que cria mais elementos próximos a um ponto especificado.

Parâmetros

baseType:::

Nx1:::

Nx2:::

a::Tuple:

b::Tuple:

ponto_foco::Tuple:

Retorno

Exemplo

Fease.monta_u_aproximada — Method

monta_u_aproximada(c::Vector{Float64}, malha::Malha)

Dada a solução obtida através do sistema linear e a malha sob a qual foi solucionada, retorna uma função que aproxima a solução analítica do problema.

Parâmetros

c::Vector{Float64}: O vetor contendo a solução do sistema linear.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

uh::function: Uma função que aproxima a solução analítica do problema.

Exemplo

Fease.mudanca_variavel_xξ — Method

mudanca_variavel_xξ(Xᵉ::Vector, Φξ::Vector, n_dim::Int)

Mapeia as coordenadas físicas x = (x₁, x₂, xᵢ...) a partir de ξ = (ξ₁, ξ₂, ξᵢ...)

Parâmetros

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

Φξ::Vector: Vetor com o valor das funções locais Φ avaliadas no ponto de Gauss ξ = (ξ₁, ξ₂, ξᵢ...).

n_dim::Int: Número de dimensões do espaço.

Retorno

x::Vector: A coordenada física referente ao ponto de Gauss ξ.

Exemplo

Fease.quadratura_gauss — Method

quadratura_gauss(npg::Int, n_dim::Int)

Define os npg pontos de Gauss-Legendre P e os pesos associados W. Cada um dos pontos em P são n-uplas dimensionadas por n_dim. Os pesos associados W funcionam analogamente.

Parâmetros

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

Fease.quadratura_ϕ — Method

quadratura_ϕ(base, npg::Int, n_dim::Int)

Avalia cada uma das funções ϕ em cada um dos npg pontos de Gauss-Legendre e armazena o resultado em uma matriz ϕP: npg x n_funcs.

Parâmetros

n_funcs::Int: Número de funções base.

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

ϕP::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

Fease.quadratura_∇ϕ — Method

quadratura_∇ϕ(base, npg::Int, n_dim::Int)

Avalia cada uma das funções ∂ϕ, para cada uma das n_dim dimensões, em cada um dos npg pontos de Gauss-Legendre e armazena o resultado em uma n-upla ∇ϕP de matrizes ∂ϕP: npg x n_funcs.

Parâmetros

n_funcs::Int: Número de funções base.

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

∇ϕP::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

Fease.solve_sys — Method

solve_sys(f::function, malha::Malha, pseudo_a::function)

Monta e soluciona o sistema linear KC = F, considerando a referência ao operador bilinear a(u,v) obtido na formulação fraca.

Parâmetros

run_values::RunValues: Estrutura que contém parâmetros do problema.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

C::Vector{Float64}: O vetor contendo a solução do sistema linear.

Exemplo

Fease.solve_sys_poisson — Method

solve_sys_poisson(run_values::RunValues, malha::Malha)

Monta e soluciona o sistema linear KC = F, considerando o problema com equação de Poisson.

Parâmetros

run_values::RunValues: Estrutura que contém parâmetros do problema.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

C::Vector{Float64}: O vetor contendo a solução do sistema linear.

Exemplo