Referências

Conteúdo

Referências

Conteúdo

Índice

Índice

Fease.Malha

Fease.RunValues

Fease.TermosEquacao

Fease.convergence_test!

Fease.elem_coords

Fease.jacobiano

Fease.malha2D_adiciona_ruido

Fease.montaEQ_geral

Fease.montaF_geral

Fease.montaFᵉ_geral!

Fease.montaK_geral

Fease.montaKᵉ_geral!

Fease.montaLG_geral

Fease.monta_LG_2D

Fease.monta_malha_1D_uniforme

Fease.monta_malha_2D_foco

Fease.monta_malha_2D_uniforme

Fease.monta_u_aproximada

Fease.mudanca_variavel_xξ

Fease.plot_malha_2D

Fease.plot_solucao_aproximada

Fease.quadratura_gauss

Fease.quadratura_ϕ

Fease.quadratura_∇ϕ

Fease.solve_sys

Fease.solve_sys_poisson

Fease.ϕ_1D

Fease.ϕ_2D

Fease.ϕ_geral

Fease.∂ϕ_∂ξ₁

Fease.∂ϕ_∂ξ₂

Fease.∇ϕ_1D

Fease.∇ϕ_2D

Fease.∇ϕ_geral

Fease.Malha — Type

struct Mesh

Valores que definem uma malha.

Campos

base::Base: Informações sobre o tipo da base de funções interpoladoras

ne::Int64: Número de elementos totais

neq::Int64: Número de equações

coords::Tuple: N-upla com as coordenadas (X₁, X₂, Xᵢ...) dos nós da malha

dx::: n-upla contendo o intervalo entre os nós, para cada eixo (se uniforme)

EQ::Vector{Int64}: Vetor com as reenumerações das funções globais ϕ

LG::Matrix{Int64}: Matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ

EQoLG::Matrix{Int64}: Matriz composta entre a EQ e a LG

a::Tuple: Coordenada do início do intervalo

b::Tuple: Coordenada do final do intervalo

n_dim::Int64: Número de dimensões da malha

Nx::Tuple: Número de subdivisões da malha para cada eixo

Fease.RunValues — Type

struct RunValues

Valores que definem parâmetros de um problema.

Campos

α::Float64: Coeficiente α

β::Float64: Coeficiente β

γ::Float64: Coeficiente γ

f::Function: Função que represnenta o lado direito da equação do problema.

u::Function: Função que representa a solução analítica da equação do problema.

Fease.TermosEquacao — Type

struct EquationTerms

Referências a termos da equação.

Campos

u::: Representa um dos pares do operador linear a(u,v).

v::: Representa um dos pares do operador linear a(u,v).

∇u::: Representa o gradiente de um dos pares do operador linear a(u,v).

∇v::: Representa o gradiente de um dos pares do operador linear a(u,v).

x::: Termo que permite representar um termo qualquer da equação que dependa de valores no domínio.

Fease.convergence_test! — Method

convergence_test!(E, NE, n_dims, monta_malha, pseudo_a, f, u)

Realiza o estudo da convergência do erro para a solução aproximada da equação com solução exata u.

Parâmetros

E::Vetor{Float64}: O vetor onde serão inseridos os erros caculados para cada um das quantidades de elementos finitos.

NE::Vetor{Float64}: O vetor que contém as cada uma das quantidades de elementos finitos.

n_dim::Int: Número de dimensões do espaço.

monta_malha::Function: Uma função que recebe os números de sub-intervalos da malha e a retorna montada.

pseudo_a: A referência ao operador bilinear a(u,v).

f::Function: Função que representa o lado direito da equação do problema.

u::Function: A solução analítica da equação.

Retorno

Altera o vetor E.

Exemplo

  # Define o tipo da base de funções interpoladoras do subespaço aproximado Vₘ.
  baseType = BaseTypes.linearLagrange

  # Define extremos do intervalo da malha.
  a = (0.0, 0.0)
  b = (1.0, 1.0)

  # Define os parâmetros da equação a ser resolvida.
  α = 1.0
  β = 1.0
  f = (x₁, x₂) -> (2 * α * π^2 + β) * sin(π * x₁) * sin(π * x₂)
  u = (x₁, x₂) -> sin(π * x₁) * sin(π * x₂)

  # Define o pseudo operador linear a(u,v).
  function pseudo_a(termos_equacao)
    (; ∇u, ∇v, u, v) = termos_equacao

    return β * dot(u, v) + α * dot(∇u, ∇v)
  end

  # Define a função que constrói uma malha que depende só do número de sub-intervalos nos dois eixos.
  function monta_malha(NX)
    return monta_malha_2D_uniforme(baseType, NX..., a, b)
  end

  # Estudo será realizado com número de elementos em potências de 2.
  errsize = 7
  NE = 2 .^ [2:1:errsize;]
  H = 1 ./ NE
  E = zeros(length(NE))

  # Calcula o erro para cada uma das quantidades de elementos finitos.
  convergence_test!(E, NE, 2, monta_malha, pseudo_a, f, u)

Fease.elem_coords — Method

elem_coords(malha::Malha, e::Int)

Encontra as coordenadas dos nós do elemento finito e e os retorna, junto a EQoLG referente.

Parâmetros

malha::Malha: A malha sob a qual está sendo solucionado o problema.

e::Int: O identificador numérico do elemento finito Ωᵉ.

Retorno

eqs_idx::: EQoLG para o elemento e passado como parâmetro

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

Exemplo

Fease.jacobiano — Method

jacobian(n_dim::Int, Xᵉ::Vector, ∇Φξ::Tuple{Matrix{Float64}}, ξ::Int)

Calcula a matriz jacobiana, assim como seu determinante, no domínio do elemento determinado pelas coordenadas em Xᵉ.

Parâmetros

n_dim::Int: Número de dimensões do espaço.

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

∇Φξ::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

ξ::Int: Representa o número do ponto de Gauss.

Retorno

M::Matrix{Float64}: Matriz jacobiana.

detJ::Float64: O determinante da matriz jacobiana

Exemplo

Fease.malha2D_adiciona_ruido — Method

monta_malha_2D_uniforme(baseType, Nx1, Nx2, a::Tuple, b::Tuple)::Malha

Função que constrói uma malha 2D uniforme.

Parâmetros

baseType::: O tipo da base das funções interpoladoras.

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

Nx2::Int64: Número de subdivisões da malha no segundo eixo.

a::Tuple: Coordenada do início do intervalo

b::Tuple: Coordenada do final do intervalo

Retorno

malha::Malha: Uma malha 2D uniforme dentro do intervalo [a,b].

Exemplo

Fease.montaEQ_geral — Function

montaEQ_geral(Nx1::Int64, Nx2::Int64 = 0)

Função que monta o vetor com as reenumerações das funções globais ϕ.

Parâmetros

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

Nx2::Int64 = 0: Número de subdivisões da malha no segundo eixo.

Retorno

EQ::Vector{Int64}: Vetor com as reenumerações das funções globais ϕ.

Exemplo

Fease.montaF_geral — Method

montaF_geral(f::Function, malha::Malha)

Monta o vetor F global do sistema linear do problema.

Parâmetros

f::function: Função que representa o lado direito da equação do problema.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

F::Vector{Float64}: O vetor F global do sistema linear do problema.

Exemplo

Fease.montaFᵉ_geral! — Method

montaFᵉ_geral!(Fᵉ, f, Xᵉ, P, W, ϕξ, ∇ϕξ, n_dim)

Monta o vetor local Fᵉ referente à contribuição do elemento e.

Parâmetros

Fᵉ::Vector{Float64}: Vetor local Fᵉ referente à contribuição do elemento e.

f::function: Funçãp que represnenta o lado direito da equação do problema.

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

ϕξ::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

∇ϕξ::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço.

Retorno

Altera Fᵉ.

Exemplo

Fease.montaK_geral — Method

montaK_geral(malha::Malha, pseudo_a)

Monta a matriz K global do sistema linear do problema.

Parâmetros

malha::Malha: A malha sob a qual está sendo solucionado o problema.

pseudo_a::: A referência ao operador bilinear a(u,v).

Retorno

K::Matrix{Float64}: A matriz K global do sistema linear do problema.

Exemplo

Fease.montaKᵉ_geral! — Method

montaKᵉ_geral!(Kᵉ, Xᵉ, P, W, Φξ, ∇Φξ, n_dim, pseudo_a)

Monta a matriz local Kᵉ referente à contribuição do elemento e.

Parâmetros

Kᵉ::Matrix{Float64}: Matriz local Kᵉ referente à contribuição do elemento e.

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Φξ::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

∇Φξ::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço.

pseudo_a: A referência ao operador bilinear a(u,v).

Retorno

Altera Kᵉ.

Exemplo

Fease.montaLG_geral — Function

montaLG_geral(Nx1::Int64, Nx2::Int64 = 0)

Função que monta a matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ.

Parâmetros

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

Nx2::Int64 = 0: Número de subdivisões da malha no segundo eixo.

Retorno

LG::Matrix{Int64}: Matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ.

Exemplo

Fease.monta_LG_2D — Method

monta_LG_2D(Nx1::Int64, Nx2::Int64)::Matrix{Int64}

Função que monta a matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ.

Parâmetros

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

Nx2::Int64: Número de subdivisões da malha no segundo eixo.

Retorno

LG::Matrix{Int64}: Matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ.

Exemplo

Fease.monta_malha_1D_uniforme — Method

monta_malha_1D_uniforme(baseType, Nx1, a, b)::Malha

Função que constrói uma malha 1D uniforme.

Parâmetros

baseType::: O tipo da base das funções interpoladoras.

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

a::Tuple: Coordenada do início do intervalo uniforme.

b::Tuple: Coordenada do final do intervalo uniforme.

Retorno

malha::Malha: Uma malha 1D uniforme dentro do intervalo [a,b].

Exemplo

Fease.monta_malha_2D_foco — Function

monta_malha_2D_foco(
baseType, Nx1, Nx2, a::Tuple, b::Tuple, ponto_foco::Tuple, precisao::Int64)::Malha

Função que constrói uma malha 2D que cria mais elementos próximos a um ponto especificado.

Parâmetros

baseType::: O tipo da base das funções interpoladoras.

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

Nx2::Int64: Número de subdivisões da malha no segundo eixo.

a::Tuple: Coordenada do início do intervalo

b::Tuple: Coordenada do final do intervalo

ponto_foco::Tuple: O ponto na malha onde elementos serão adicionados em volta.

precisao::Int64 = 1: A quantidade de iterações de incremento de intervalos em torno do ponto_foco

Retorno

malha::Malha: Uma malha 2D com concentração de elementos em ponto_foco.

Exemplo

Fease.monta_malha_2D_uniforme — Method

monta_malha_2D_uniforme(baseType, Nx1, Nx2, a::Tuple, b::Tuple)::Malha

Função que constrói uma malha 2D uniforme.

Parâmetros

baseType::: O tipo da base das funções interpoladoras.

Nx1::Int64: Número de subdivisões da malha no primeiro eixo.

Nx2::Int64: Número de subdivisões da malha no segundo eixo.

a::Tuple: Coordenada do início do intervalo

b::Tuple: Coordenada do final do intervalo

Retorno

malha::Malha: Uma malha 2D uniforme dentro do intervalo [a,b].

Exemplo

Fease.monta_u_aproximada — Method

monta_u_aproximada(c::Vector{Float64}, malha::Malha)

Dada a solução obtida através do sistema linear e a malha sob a qual foi solucionada, retorna uma função que aproxima a solução analítica do problema.

Parâmetros

c::Vector{Float64}: O vetor contendo a solução do sistema linear.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

uh::function: Uma função que aproxima a solução analítica do problema.

Exemplo

Fease.mudanca_variavel_xξ — Method

mudanca_variavel_xξ(Xᵉ::Vector, Φξ::Vector, n_dim::Int)

Mapeia as coordenadas físicas x = (x₁, x₂, xᵢ...) a partir de ξ = (ξ₁, ξ₂, ξᵢ...)

Parâmetros

Xᵉ::Vector: Coordenadas dos vértices do elemento finito Ωᵉ. Cada entrada do vetor representa um eixo.

Φξ::Vector: Vetor com o valor das funções locais Φ avaliadas no ponto de Gauss ξ = (ξ₁, ξ₂, ξᵢ...).

n_dim::Int: Número de dimensões do espaço.

Retorno

x::Vector: A coordenada física referente ao ponto de Gauss ξ.

Exemplo

Fease.plot_malha_2D — Method

plot_malha_2D(malha::Malha)

Plota o gráfico que representa a malha.

Parâmetros

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

Dá display no gráfico plotado.

Exemplo

Fease.plot_solucao_aproximada — Function

plot_solucao_aproximada(c::Array, malha::Malha, gif::Bool = false)

A partir da solução do sistema e a malha, plota a solução aproximada. No caso do espaço ser bidimensional é possível atribuir gif = true para que o plot seja um gif animado com visão 360⁰.

Parâmetros

c::Array: O vetor contendo a solução do sistema linear.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

gif::Bool = false: Flag para plotar o gráfico como um gif animado.

Retorno

Não há retorno, porém dá display no gráfico plotado.

Exemplo

Fease.quadratura_gauss — Method

quadratura_gauss(npg::Int, n_dim::Int)

Define os npg pontos de Gauss-Legendre P e os pesos associados W. Cada um dos pontos em P são n-uplas dimensionadas por n_dim. Os pesos associados W funcionam analogamente.

Parâmetros

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

Fease.quadratura_ϕ — Method

quadratura_ϕ(base, npg::Int, n_dim::Int)

Avalia cada uma das funções ϕ em cada um dos npg pontos de Gauss-Legendre e armazena o resultado em uma matriz ϕP: npg x n_funcs.

Parâmetros

n_funcs::Int: Número de funções base.

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

ϕP::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

Fease.quadratura_∇ϕ — Method

quadratura_∇ϕ(base, npg::Int, n_dim::Int)

Avalia cada uma das funções ∂ϕ, para cada uma das n_dim dimensões, em cada um dos npg pontos de Gauss-Legendre e armazena o resultado em uma n-upla ∇ϕP de matrizes ∂ϕP: npg x n_funcs.

Parâmetros

n_funcs::Int: Número de funções base.

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

∇ϕP::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

Fease.solve_sys — Method

solve_sys(f::function, malha::Malha, pseudo_a::function)

Monta e soluciona o sistema linear KC = F, considerando a referência ao operador bilinear a(u,v) obtido na formulação fraca.

Parâmetros

run_values::RunValues: Estrutura que contém parâmetros do problema.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

C::Vector{Float64}: O vetor contendo a solução do sistema linear.

Exemplo

Fease.solve_sys_poisson — Method

solve_sys_poisson(run_values::RunValues, malha::Malha)

Monta e soluciona o sistema linear KC = F, considerando o problema com equação de Poisson.

Parâmetros

run_values::RunValues: Estrutura que contém parâmetros do problema.

malha::Malha: A malha sob a qual está sendo solucionado o problema.

Retorno

C::Vector{Float64}: O vetor contendo a solução do sistema linear.

Exemplo

Fease.ϕ_1D — Method

ϕ_1D(P...)

Calcula os valores das funções interpoladoras ϕ no ponto P, unidimensional.

Parâmetros

P...: Ponto onde as funções interpoladoras serão calculadas.

Retorno

ϕ::Vector{Float64}: Vetor com os valores das funções interpoladoras ϕ no ponto P.

Exemplo

Fease.ϕ_2D — Method

ϕ_2D(ξ₁::Float64, ξ₂::Float64)::Vector{Float64}

Calcula os valores das funções interpoladoras ϕ no ponto P, unidimensional.

Parâmetros

ξ₁::Float64: Valor no primeiro eixo do ponto onde as funções interpoladoras serão calculadas.

ξ₂::Float64: Valor no segundo eixo do ponto onde as funções interpoladoras serão calculadas.

Retorno

ϕ::Vector{Float64}: Vetor com os valores das funções interpoladoras ϕ no ponto P.

Exemplo

Fease.ϕ_geral — Method

ϕ_geral(P...)

Calcula os valores das funções interpoladoras ϕ no ponto P.

Parâmetros

P...: Ponto onde as funções interpoladoras serão calculadas.

Retorno

ϕ::Vector{Float64}: Vetor com os valores das funções interpoladoras ϕ no ponto P.

Exemplo

Fease.∂ϕ_∂ξ₁ — Method

∂ϕ_∂ξ₁(ξ₂::Float64)::Vector{Float64}

Definição das ∂ϕ_∂ξ₁ lagrange linear para o caso 2D

Parâmetros

ξ₂::Float64: Valor no segundo eixo do ponto onde as funções interpoladoras serão calculadas.

Retorno

∂ϕ_∂ξ₁::Vector{Float64}: Vetor com os valores das derivadas das funções interpoladoras ϕ no ponto P, no primeiro eixo.

Exemplo

Fease.∂ϕ_∂ξ₂ — Method

∂ϕ_∂ξ₂(ξ₁::Float64)::Vector{Float64}

Definição das ∂ϕ_∂ξ₂ lagrange linear para o caso 2D

Parâmetros

ξ₁::Float64: Valor no primeiro eixo do ponto onde as funções interpoladoras serão calculadas.

Retorno

∂ϕ_∂ξ₂::Vector{Float64}: Vetor com os valores das derivadas das funções interpoladoras ϕ no ponto P, no segundo eixo.

Exemplo

Fease.∇ϕ_1D — Method

∇ϕ_1D(P...)

Calcula os valores das derivadas das funções interpoladoras ϕ no ponto P.

Parâmetros

P...: Ponto onde as funções interpoladoras serão calculadas.

Retorno

∇ϕ::Vector{Float64}: Vetor com os valores das derivadas das funções interpoladoras ϕ no ponto P.

Exemplo

Fease.∇ϕ_2D — Method

∇ϕ_2D(ξ₁::Float64, ξ₂::Float64)

Calcula os valores do gradiente das funções interpoladoras ϕ no ponto P.

Parâmetros

ξ₁::Float64: Valor no primeiro eixo do ponto onde as funções interpoladoras serão calculadas.

ξ₂::Float64: Valor no segundo eixo do ponto onde as funções interpoladoras serão calculadas.

Retorno

∇ϕ::Tuple: Tupla com os valores do gradiente das funções interpoladoras ϕ no ponto P.

Exemplo

Fease.∇ϕ_geral — Method

∇ϕ_geral(P...)

Calcula os valores do gradiente das funções interpoladoras ϕ no ponto P.

Parâmetros

P...: Ponto onde as funções interpoladoras serão calculadas.

Retorno

∇ϕ::Tuple: Tupla com os valores do gradiente das funções interpoladoras ϕ no ponto P.

Exemplo