Referências

Conteúdo

Referências

Conteúdo

Índice

Índice

MyProject.Malha

MyProject.elem_coords

MyProject.montaF_geral

MyProject.montaFᵉ_geral!

MyProject.montaK_geral

MyProject.montaKᵉ_geral!

MyProject.mudanca_variavel_xξ

MyProject.quadratura_gauss

MyProject.quadratura_ϕ

MyProject.quadratura_∇ϕ

MyProject.solveSys_geral

MyProject.Malha — Type

struct Mesh

Valores que definem uma malha.

Campos

base::Base: Informações sobre o tipo da base de funções interpoladoras

ne::Int64: Número de elementos totais

neq::Int64: Número de equações

coords::Tuple: N-upla com as coordenadas (X₁, X₂, Xᵢ...) dos nós da malha

dx: n-upla contendo o intervalo entre os nós, para cada eixo (se uniforme)

EQ: Vetor com as reenumerações das funções globais ϕ

LG: Matriz de conectividade local/global (LG). Relaciona a numeração local e global das funções ϕ

EQoLG::Matrix{Int64}: Matriz composta entre a EQ e a LG

a: Coordenada do início do intervalo uniforme

b: Coordenada do final do intervalo uniforme

n_dim: Número de dimensões da malha

Nx: Número de subdivisões da malha para cada eixo

MyProject.elem_coords — Method

elem_coords(malha::Malha, e::Int)

Descrição.

Parâmetros

malha::Malha:

e::Int:

Retorno

eqs_idx:::

Xᵉ:::

Exemplo

MyProject.montaF_geral — Method

montaF_geral(run_values::RunValues, malha::Malha)

Descrição.

Parâmetros

run_values::RunValues:

malha::Malha:

Retorno

F::Vector{Float64}:

Exemplo

MyProject.montaFᵉ_geral! — Method

montaFᵉ_geral!(Fᵉ, f, Xᵉ, P, W, ϕξ, ∇ϕξ, n_dim)

Descrição.

Parâmetros

Fᵉ:::

f:::

Xᵉ:::

P:::

W:::

ϕξ:::

∇ϕξ:::

n_dim:::

Retorno

Altera Fᵉ.

Exemplo

MyProject.montaK_geral — Method

montaK_geral(run_values::RunValues, malha::Malha)

Descrição.

Parâmetros

run_values::RunValues:

malha::Malha:

Retorno

K::Matrix{Float64}:

Exemplo

MyProject.montaKᵉ_geral! — Method

montaKᵉ_geral!(Kᵉ, Xᵉ, P, W, Φξ, ∇Φξ, n_dim, dx, run_values::RunValues)

Descrição.

Parâmetros

Kᵉ:::

Xᵉ:::

P:::

W:::

Φξ:::

∇Φξ:::

n_dim:::

run_values::RunValues:

Retorno

Altera Kᵉ.

Exemplo

MyProject.mudanca_variavel_xξ — Method

mudanca_variavel_xξ(Xᵉ, Φξ, n_dim)

Descrição.

Parâmetros

Xᵉ:::

Φξ:::

n_dim::Int: Número de dimensões do espaço.

Retorno

x:::

Exemplo

MyProject.quadratura_gauss — Method

quadratura_gauss(npg::Int, n_dim::Int)

Define os npg pontos de Gauss-Legendre P e os pesos associados W. Cada um dos pontos em P são n-uplas dimensionadas por n_dim. Os pesos associados W funcionam analogamente.

Parâmetros

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

MyProject.quadratura_ϕ — Method

quadratura_ϕ(base, npg::Int, n_dim::Int)

Avalia cada uma das funções ϕ em cada um dos npg pontos de Gauss-Legendre e armazena o resultado em uma matriz ϕP: npg x n_funcs.

Parâmetros

n_funcs::Int: Número de funções base.

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

ϕP::Matrix{Float64}: Matriz npg x n_funcs com a avaliaçao das ϕ nos pontos de Gauss-Legendre.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

MyProject.quadratura_∇ϕ — Method

quadratura_∇ϕ(base, npg::Int, n_dim::Int)

Avalia cada uma das funções ∂ϕ, para cada uma das n_dim dimensões, em cada um dos npg pontos de Gauss-Legendre e armazena o resultado em uma n-upla ∇ϕP de matrizes ∂ϕP: npg x n_funcs.

Parâmetros

n_funcs::Int: Número de funções base.

npg::Int: Número de pontos de integração de Gauss-Legendre.

n_dim::Int: Número de dimensões do espaço. Tamanho dos pontos de Gauss-Legendre.

Retorno

∇ϕP::Tuple{Matrix{Float64}}: N-upla de matrizes npg x n_funcs com a avaliaçao das ∂ϕ nos pontos de Gauss-Legendre.

P::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um ponto de Gauss-Legendre dimensionado por n_dim.

W::Vector{Tuple}: Vetor de n-uplas, onde cada n-upla é um conjunto ordenado de pesos de Gauss-Legendre associados a cada um dos pontos em P.

Exemplo

MyProject.solveSys_geral — Method

solveSys_geral(run_values::RunValues, malha::Malha)

Monta e soluciona o sistema linear KC = F.

Parâmetros

run_values::RunValues:

malha::Malha:

Retorno

C::Vector{Float64}:

Exemplo